Edit sp eliminar usuario y delete cascade

+++ StoredProcedures/Usuario/sp_eliminar_usuario.sql
@@ -1,28 +1,36 @@
 -- SP ELIMINAR USUARIO
-create or replace procedure sp_eliminar_usuario(rut_u in varchar, removed out boolean)
-is begin
-    delete from usuario
+create or replace procedure sp_eliminar_usuario(rut_u in varchar, removed out number)
+is 
+    user_c number;
+begin
+    select count(*) into user_c
+    from usuario
     where numrut = rut_u;
 
-    removed := true;
+    if user_c = 0 then
+        removed := 0;
+    else
+        delete from usuario
+        where numrut = rut_u;
+        removed := 1;
+        commit;
+    end if;
 exception
     when others then
-        dbms_output.put_line('Se deben eliminar los registros de las tablas dependientes');
-        removed := false;
+        removed := 0;
 end;
 
 /
--- todo: se deben eliminar registros de tablas dependientes
 
 declare
     rut varchar2(8);
-    removed boolean;
+    removed number;
 begin
     rut := '15624578';
-    removed := false;
+    removed := 0;
     sp_eliminar_usuario(rut, removed);
     
-    if removed = true then
+    if removed = 1 then
         dbms_output.put_line('Usuario Eliminado.');
     else
         dbms_output.put_line('Error al eliminar usuasrio');
